rename AuxOps to SupportOps
