+++ tla-bmcmt/src/main/scala/at/forsyte/apalache/tla/bmcmt/rules/FunExceptRuleWithArrays.scala
@@ -2,7 +2,7 @@ package at.forsyte.apalache.tla.bmcmt.rules
 
 import at.forsyte.apalache.tla.bmcmt._
 import at.forsyte.apalache.tla.bmcmt.arena.PtrUtil
-import at.forsyte.apalache.tla.bmcmt.rules.support.AuxOps.constrainRelationArgs
+import at.forsyte.apalache.tla.bmcmt.rules.support.SupportOps.constrainRelationArgs
 import at.forsyte.apalache.tla.types.{BuilderUT => BuilderT, tlaU => tla}
 
 /**
