+++ tla-bmcmt/src/main/scala/at/forsyte/apalache/tla/bmcmt/rules/support/CherryPick.scala
@@ -3,7 +3,7 @@ package at.forsyte.apalache.tla.bmcmt.rules.support
 import at.forsyte.apalache.io.config.SMTEncoding
 import at.forsyte.apalache.tla.bmcmt._
 import at.forsyte.apalache.tla.bmcmt.arena.PtrUtil
-import at.forsyte.apalache.tla.bmcmt.rules.support.AuxOps._
+import at.forsyte.apalache.tla.bmcmt.rules.support.SupportOps._
 import at.forsyte.apalache.tla.bmcmt.types._
 import at.forsyte.apalache.tla.lir._
 import at.forsyte.apalache.tla.typecomp._
