+++ tla-bmcmt/src/main/scala/at/forsyte/apalache/tla/bmcmt/rules/support/SupportOps.scala
@@ -12,7 +12,7 @@ import at.forsyte.apalache.tla.types.{BuilderT, tla}
  * @author
  *   Rodrigo Otoni
  */
-object AuxOps {
+object SupportOps {
   private lazy val simplifier = new ConstSimplifierForSmt
 
   /**
